04 auth

+++ services/contactsServices.js
@@ -1,27 +1,27 @@
 import { Contact } from "../models/contact.js";
 
-const listContacts = async () => {
-  return Contact.find();
+const listContacts = async (owner, skip, limit) => {
+  return Contact.find({ owner }, "", { skip, limit });
 };
 
-const getContactById = async (id) => {
-  return Contact.findOne({ _id: id });
+const getContactById = async (id, owner) => {
+  return Contact.findOne({ _id: id, owner });
 };
 
-const removeContact = async (id) => {
-  return Contact.findOneAndDelete({ _id: id });
+const removeContact = async (id, owner) => {
+  return Contact.findOneAndDelete({ _id: id, owner });
 };
 
-const addContact = async (name, email, phone) => {
-  return Contact.create({ name, email, phone });
+const addContact = async (body) => {
+  return Contact.create(body);
 };
 
-const updateContact = async (id, data) => {
-  return Contact.findByIdAndUpdate({ _id: id }, data, { new: true });
+const updateContact = async (id, data, owner) => {
+  return Contact.findOneAndDelete({ _id: id, owner }, data, { new: true });
 };
-const updateStatusContact = async (id, status) => {
-  return Contact.findByIdAndUpdate(
-    { _id: id },
+const updateStatusContact = async (id, status, owner) => {
+  return Contact.findOneAndDelete(
+    { _id: id, owner },
     { favorite: status },
     { new: true }
   );
